Workaround another installer bug
When the user tries to install an incompatible app, certain Android versions don't send us the error broadcast. So we think the user simply clicked outside the confirmation dialog and ask again, leading to a UI dialog loop. As a simple workaround we use a counter and bail out after showing the dialog three times.